hmm.. the disk_ and tape_ backup scripts don't really need to be different scripts, the actual fundamental difference is that tape_ needs the backup host to be ssh accessable to the clients, while disk_ just needs the clients to be accessable to the server.
the way tape_ does it is better, we get better logs for one thing
( because the way disk_ does it is to stream the output of tar
back to the backup shell which started SSH, while tape_ tells
the backup client to use the backup server's tape dev via ssh ).

i will merge these into one eventually, i need to figure out the
security issue.. i don't think all clients should have direct
access to the backup server's tape drive ( although this is
a more "traditional" unix way to do it ), it's just not neccessary.